feat(activities): add empty state & refactor
