+++ app/src/main/java/com/teobaranga/monica/contacts/detail/ContactDetailViewModel.kt
@@ -3,7 +3,7 @@ package com.teobaranga.monica.contacts.detail
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.teobaranga.monica.contacts.data.ContactRepository
-import com.teobaranga.monica.contacts.detail.ui.ContactInfoActivitiesSection
+import com.teobaranga.monica.contacts.detail.activities.ui.ContactInfoActivitiesSection
 import com.teobaranga.monica.contacts.detail.ui.ContactInfoContactSection
 import com.teobaranga.monica.contacts.detail.ui.ContactInfoPersonalSection
 import com.teobaranga.monica.contacts.detail.ui.ContactInfoRelationshipsSection
