+++ app/src/main/java/com/teobaranga/monica/contacts/detail/activities/ui/ContactActivitiesUiState.kt
@@ -1,9 +1,10 @@
-package com.teobaranga.monica.contacts.detail.activities
+package com.teobaranga.monica.contacts.detail.activities.ui
 
 import androidx.compose.runtime.Immutable
 
 sealed interface ContactActivitiesUiState {
     data object Loading : ContactActivitiesUiState
+    data object Empty : ContactActivitiesUiState
 
     @Immutable
     data class Loaded(
