+++ app/src/main/java/com/teobaranga/monica/contacts/detail/ui/ContactInfoSection.kt
@@ -3,7 +3,7 @@ package com.teobaranga.monica.contacts.detail.ui
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 
-sealed interface ContactInfoSection {
+interface ContactInfoSection {
     val title: String
 
     @Composable
